Change steamguard-cli command to steamguard in example (#189)
A very minor pull request for a small (I believe) mistake in the
`README.md`. The command for generating and copying a new code to the
clipboard used `steamguard-cli` as command which took me a bit off guard
and made me question if I installed steamguard-cli correctly.

This PR changes it to `steamguard`.