feat(confirmations): use smaller struct for confirmer methods (#475)
Add a `ConfirmationId<'a>` struct that contains only the `id` and
`nonce` fields used for approvals and refusals of confirmations.

For reverse compatibility I changed the methods to use `impl
Into<ConfirmationId<'a>>` parameters, which are compatible with both
`&Confirmation` and the introduced `ConfirmationId` struct. Let me know
if there's a better way to do this 😃

Closes #474